requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-modeltranslation==0.18.11
idna==3.6
pillow==10.3.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
